Fixat lite med Category så man kan tabort etc

+++ BL/Controllers/CategoryController.cs
@@ -4,14 +4,16 @@ using DAL;
 using Models;
 using DAL.Repositories;
 
+
 namespace BL.Controllers
 {
     public class CategoryController
     {
         ICategoryRepository<Category> categoryRepository;
-
+        Validering validator;
         public CategoryController()
         {
+            validator = new Validering();
             categoryRepository = new CategoryRepository();
         }
         public void createCategory(string name)
@@ -25,8 +27,15 @@ namespace BL.Controllers
         }
         public void deleteCategory(string category)
         {
-            int i = categoryRepository.GetIndexOfName(category);
-            categoryRepository.Delete(i);
+            if (validator.TextEmpty(category))
+            {
+                int i = categoryRepository.GetIndexOfName(category);
+                categoryRepository.Delete(i);
+            }
+            else
+            {
+                Console.WriteLine("No category");
+            }
         }
 
         public void updatCategory(Category oldCat, string newCategoryName)
